requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
apache-airflow==2.5.0
contourpy==1.0.6
    # via matplotlib
cycler==0.11.0
    # via matplotlib
fonttools==4.38.0
    # via matplotlib
kiwisolver==1.4.4
    # via matplotlib
matplotlib==3.6.2
    # via
    #   -r requirements.in
    #   seaborn
numpy==1.23.5
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
    #   scipy
    #   seaborn
packaging==21.3
    # via matplotlib
pandas==1.5.1
    # via
    #   -r requirements.in
    #   seaborn
pillow==9.3.0
    # via matplotlib
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2022.6
    # via pandas
scipy==1.9.3
    # via seaborn
seaborn==0.10.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil